- removed deprecated evaluation REST endpoints and other code - merged EvaluationRESTService and Evaluation2RESTService - renamed some misleading classes - resolves #6276
